Simplified node API
